Move value interfaces into a grouping class `Element`
