+++ server/src/test/java/io/spine/server/tuple/TupleShould.java
@@ -23,8 +23,8 @@ package io.spine.server.tuple;
 import com.google.protobuf.BoolValue;
 import com.google.protobuf.Empty;
 import com.google.protobuf.Message;
-import io.spine.server.tuple.Tuple.AValue;
-import io.spine.server.tuple.Tuple.BValue;
+import io.spine.server.tuple.Element.AValue;
+import io.spine.server.tuple.Element.BValue;
 import org.junit.Test;
 
 import static org.junit.Assert.assertEquals;
